Added Items.csv, battle conversion complete.
Important Note: If you open one of the CSV files for the Memoria settings in Libreoffice it is imperative that you save it as a text CSV, then open it in Notepad++ and delete the semicolons that it inserted into the lines starting with '#'. If you forget to do this, it will crash the game.