Added prison cage fight fix.
I'm unsure whether git will be able to be used for the purpose of managing steam mod files that originate from Hades Workshop. If multiple steam mod files change the same bytes file (like one of the ones committed here) it is not going to work. For now, this is the only fix that has been necessary.